# Conflicts:
#	cybexMobile/ViewControllers/Account/Account.storyboard
#	cybexMobile/ViewControllers/Account/AccountViewController.swift

+++ cybexMobile/addons/libfc/include/fc/time.hpp
@@ -52,7 +52,7 @@ namespace fc {
         static time_point from_iso_string( const fc::string& s );
 
         const microseconds& time_since_epoch()const { return elapsed; }
-        uint32_t            sec_since_epoch()const  { return elapsed.count() / 1000000; }
+        long long            sec_since_epoch()const  { return elapsed.count() / 1000000; }
         bool   operator > ( const time_point& t )const                              { return elapsed._count > t.elapsed._count; }
         bool   operator >=( const time_point& t )const                              { return elapsed._count >=t.elapsed._count; }
         bool   operator < ( const time_point& t )const                              { return elapsed._count < t.elapsed._count; }
